datasets/_scripts/curate/pyproject.toml
[project]
name = "mimyo-curate"
version = "0.1.0"
description = "Curation pipeline for mimyo face datasets — _inbox/ → raw/ → curated/ + captions"
requires-python = ">=3.11"
dependencies = [
    "pillow>=10.0",
    "imagehash>=4.3",
    "insightface>=0.7",
    "onnxruntime>=1.16",
    "requests>=2.31",
    "click>=8.1",
    "tqdm>=4.66",
    "numpy>=1.26",
    # Florence-2 / CogFlorence 캡셔너
    # transformers 4.50+ 부터 _supports_sdpa 를 모델 클래스에서 강제 →
    # Florence-2 의 custom code (trust_remote_code) 와 BC break.
    # 5.x 도 forced_bos_token_id 제거 등 BC break. 4.49 이하로 핀.
    "torch>=2.2",
    "transformers>=4.44,<4.50",
    "einops>=0.7",
    "timm>=1.0",
    "accelerate>=0.30",
]

[project.scripts]
mimyo-curate = "mimyo_curate.__main__:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mimyo_curate"]
